+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [1.4.45] - 2023-04-14
+
+### Changed
+- Failure to convert DOCX to PDF will raise an Exception.
+
+### Fixed
+- JavaScript issue with the "None of the above" checkboxes when
+  choice-specific `help` is applied to a checkbox choice.
+
 ## [1.4.44] - 2023-04-09
 
 ### Added